forcing bench_t to be a 64bit integer, this fixes overflows on some platforms -- thanks @luksan

+++ source/common/benchmarks.h
@@ -3,7 +3,7 @@
 
 struct Arguments;
 
-typedef unsigned long bench_t;
+typedef unsigned long long bench_t;
 
 typedef struct Benchmarks {
 	// Start of the total benchmarking
